# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proq API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/invite/{token}": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Preview Invite",
        "description": "What the accept screen shows before the invitee commits. Never reveals\nwhether a token is real beyond valid/invalid + a coarse reason.",
        "operationId": "preview_invite_api_invite__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invite/{token}/accept": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Accept Invite",
        "description": "Redeem an invite: create the user in the inviting org and log them in.\n\nThe email is taken from the invite (the invitee can't change who was\ninvited); only name + password come from the request.",
        "operationId": "accept_invite_api_invite__token__accept_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/team": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Get Team",
        "description": "The org's roster: current members plus still-open invitations.",
        "operationId": "get_team_api_team_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMembers"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/team/invites": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Create Invite",
        "description": "Invite a teammate by email to join the caller's organization.",
        "operationId": "create_invite_api_team_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/team/invites/{invite_id}/resend": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Resend Invite",
        "description": "Re-send a pending invitation (or, with no email provider, hand back\nthe accept link again).",
        "operationId": "resend_invite_api_team_invites__invite_id__resend_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/team/invites/{invite_id}": {
      "delete": {
        "tags": [
          "team"
        ],
        "summary": "Revoke Invite",
        "description": "Cancel a pending invitation. 404 for an unknown id or another org's.",
        "operationId": "revoke_invite_api_team_invites__invite_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "AcceptInviteRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Invite": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "invitedByUserId": {
            "type": "string",
            "title": "Invitedbyuserid"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "acceptedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptedat"
          },
          "emailed": {
            "type": "boolean",
            "title": "Emailed",
            "default": true
          },
          "emailedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emailedat"
          },
          "emailError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emailerror"
          },
          "acceptUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepturl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "status",
          "invitedByUserId"
        ],
        "title": "Invite",
        "description": "A pending/accepted/revoked invitation (never exposes the token)."
      },
      "InviteCreateRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "InviteCreateRequest"
      },
      "InvitePreview": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "organizationName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizationname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "InvitePreview",
        "description": "What the public accept screen shows before the invitee commits: which org\nthey're joining and at which email. `valid` is false for a revoked, accepted,\nexpired, or unknown token."
      },
...
      "TeamMembers": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array",
            "title": "Members"
          },
          "invites": {
            "items": {
              "$ref": "#/components/schemas/Invite"
            },
            "type": "array",
            "title": "Invites"
          }
        },
        "type": "object",
        "required": [
          "members",
          "invites"
        ],
        "title": "TeamMembers"
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "organizationId": {
            "type": "string",
            "title": "Organizationid"
          },
          "ccEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ccemail"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "company",
          "organizationId"
        ],
        "title": "User",
        "description": "Public-safe user shape (never includes the password hash)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}